Silence warning.

+++ common/Package.hpp
@@ -79,6 +79,6 @@ static const char PackageHeader[PackageHeaderSize] = { '\0', 'U', 's', 'e', 'n',
 static inline uint64_t PackageAlign( uint64_t offset ) { return ( ( offset + 7 ) / 8 ) * 8; }
 
 
-static_assert( PackageFiles == PackageFile::NUM_PACKAGE_FILE_TYPES, "Package tables mismatch." );
+static_assert( (int)PackageFiles == (int)PackageFile::NUM_PACKAGE_FILE_TYPES, "Package tables mismatch." );
 
 #endif
